Extracted test helpers.
This test framework allows to access cmake project parts,
and fake vscode parts for extension tests.
This commit moves the helper classes into a separate helper folder for
using in other test suites.
